qgrid version 1.0 from tim_shawver
I just needed
conda install -c tim_shawver qgrid
in order to get the package in Linux. This is the only way I could find to install the latest version of this package (version 1.0). Otherwise we get version 0.3.3, which is horrible.
Would you test in Windows or Mac?
There are some extra commands for Windows.

I hope I updated it correctly.